Update scripts and refactor ScriptUtils class
The install scripts for PostgreSQL, MongoDB, and Redis have been updated to ask for a sudo password at the start of the script run. This change streamlines the installation processes. Also, the ScriptUtils class has been refactored to BashRunner for better clarity of its purpose. This includes re-enabling console output for the BashRunner class.